Force static libbiomeval deps on macOS release
We can't enable this all the time because people using Homebrew on
macOS won't have all the required static dependencies. MacPorts does
not have this issue, but we'd like to support both dependency
managers.

Addtionally, zlib is only available in the dyld cache on macOS 11+
(not as a file on disk), so OpenCV must build its own version to
be found without patching OpenCV.

Regression since e7c3ba5.