Fix 392, Border cases in FDA calculation
Used to return 1.0 (highest Quality) for a maximum at the lowest frequency - half the image being black, half white.  Changed that to a value consistent with the other values returned.

Note that this only looks at sharpness of frequencies, not whether the frequencies returned are sensible ridge distances at 500 DPI.